LUCENE-9334 Consistency of field data structures
Require consistency between data-structures on a per-field basis

A field must be indexed with the same index options and data-structures across
all documents within a segment. Thus, for example, it is not allowed to have
one document in a segment where a certain field is indexed with doc values
and points, and another document where the same field is indexed only with
points. But it is allowed for a document not to have a certain field at all.

This requirement is ensured only per segment basis, thus it is possible
to have a index where two segments have different data structures
for the same field.

This patch also corrects tests where fields were indexed inconsistently
with different data structures across documents:
- Remove TestDocValuesIndexing::testDocsWithField, as it doesn't do
test anything special, what it was originally created to test