LUCENE-9838: simd version of VectorUtil.dotProduct (MH version)
This version requires no build changes or MR-JAR.

The user can opt-in to "testing" the JDK-16 incubating vector api in the
usual fashion, by passing "--add-modules jdk.incubator.vector" on the
command line.

If they do this, the dot product is accelerated, otherwise it is not.

User can also check that their setup is correct via a boolean
VectorUtil.DOTPRODUCT_VECTORIZATION_SUPPORTED.