revert most changes to BaseTokenStreamTestCase (but report out results from checkAnalysisConsistency(...))
The changes that remain don't change any of the internal workings
of the test methods, they simply expose the accumulated results
as return values so that results may be cross-checked for parity
with other analysis implementations.

This seems useful (and low-risk), but given that the rollback-based
implementation that we're using this to compare the streaming-based
impl with is not destined to be committed, it's likely that this
change can/should be rolled back as well?

Either way, we leave this slight change in place for now (for the
same reason that we leave the rollback-based impl in place for the
time being).