LUCENE-10033: Encode numeric doc values and ordinals of SORTED(_SET) doc values in blocks.
This moves doc values to an approach that is more similar to postings, where
values are grouped in blocks of 128 values that are compressed together.
Decoding a single value requires decoding the entire block that contains the
value.

Decoding of blocks is expected to be rather efficient thanks to the use of loops
that auto-vectorize. So hopefully even queries that rarely consume more than 1
value per block should remain efficient. On the other hand, queries that consume
entire blocks should be more efficient than the previous approach that would
always decode one value at a time.