LUCENE-10182: Specialize FieldSchema#assertSame for longs.
I noticed a bug in `FieldSchema#assertSame` on `docValuesGen`. Since
`docValuesGen` is a `long`, we end up calling
`assertSame(String label, Object expected, Object actual)` which checks using
object identity. In addition to the performance overhead (I only found this
because it showed up in a profile), this would be wrong for large values of
`docValuesGen` that wouldn't return cached longs.

In order to remove the trap, I replaced `Object` with `Enum` in the method
signature.