LUCENE-10154 NumericLeafComparator to define getPointValues (#364)
This patch adds getPointValues to NumericLeafComparatorsimilar how it
has getNumericDocValues.

Numeric Sort optimization with points relies on the assumption that
points and doc values record the same information, as we substitute
iterator over doc_values with one over points.

If we override getNumericDocValues it almost certainly means that whatever
PointValues NumericComparator is going to look at shouldn't be used to
skip non-competitive documents. Returning null for pointValues in this
case will force comparator NOT to use sort optimization with points,
and continue with a traditional way of iterating over doc values.