Add VarHandle variants of Packed64
Slightly modify packing tests to use the new variant(s)
Add benchmark to measure performance

The commit introduces two variants:
~ Packed64LongLong
This is the same as Packed64 however instead of using array access it
relies on VarHandle and used mainly to check the benchmark performance.
~ Packed64LongByte
A different implementation of Packed64 which reads a single long and
just a byte in case of overflow. The algorithm is slightly different
since the overflowing bits need to be computed on a different block size
which requires extra cycles.
Worse yet, it seems the memory access and operations are highly
influenced by 64-bit alignment.

Related LUCENE-10205