LUCENE-10010: don't determinize/minimize in RegExp
Today, RegExp calls minimize() at every parsing step. There is little
point to making an NFA execution when it is doing this.

Moreover, some minimize() calls are missing, and in fact in rare cases
it can already return an NFA today (for certain syntax)

Instead, RegExp parsing should do none of this, instead it may return a
DFA or NFA. NOTE: many simple regexps happen to be still returned as DFA, just
because of the algorithms in use.

Callers can decide whether to determinize or minimize. RegExp parsing should
not run in exponential time.

All src/java callsites were modified to call minimize(), to prevent any
performance problems. minimize() seems unnecessary, but let's approach
removing minimization as a separate PR. src/test was fixed to just use
determinize() in preparation for this.