LUCENE-10311: add long method, that forwards to int method and truncates
Confusion happens because grow(numAdds) reserves space for you to call
add() up to numAdds times.

When numAdds exceeds a "threshold" (maxdoc >> 8), we really don't care
about big numbers at all: we'll switch to a FixedBitSet(maxDoc) with
fixed sized storage, bounded only by maxDoc.

But we can just add a one-liner grow(long) that doesn't require the
caller to understand any of this, and hide it via implementation detail.