LUCENE-10486: Avoid unnecessary overhead in TopScoreDoc and TopField collector manager
The createSharedManager static methods exposed by TopFieldCollector and TopScoreDocCollector rely unconditionally on a shared global counter as well as a shared max score accumulator. These can cause needless overhead when the search is not executed concurrently. This commit adds a new createManager method to both collectors that creates the proper hits threshold checker and max score accumulator depending on the provided arguments, specifically the total hits threshold and whether the search is multi-threaded or not. The searcher is considered multi-threaded when its leafSlices are set (meaning the executor was provided at contruction time) and there's more than one of them. Also, when totalHitsThreshold is Integer.MAX_VALUE a special hits threshold checker is used which does no counting, and no shared max score accumulator is used, regardless of whether the searcher is multi-threaded or not.