Fixed polygon tessellation for LUCENE-10563
This issue included three polygons:

* Polygon-1 which jammed during triangulation and this PR fixes that by being more careful on selecting diagonals during the split phase
* Polygon-2 which is an invalid polygon with crossing lines, and we only add tests to assert on correct behaviour
* Polygon-3 also contains crossing lines and we add tests for asserting correct behaviour

In order to investigate the first polygon issue, we implemented a monitor interface that allows detailed information of the current state of the triangulation algorithm to be provided to external listeners for analysis.
A complete monitor allowing visualization of triangulation state has been published to another repo at https://github.com/craigtaverner/polygon-triangulator-debug.
The basic instrumenting of the triangulation algorithm for this monitor is included in the followup PR https://github.com/apache/lucene/pull/946.