LUCENE-10502: Use IndexedDISI to store docIds and DirectMonotonicWriter/Reader to handle ordToDoc (#880)
Currently vector's all docs of all fields are fully loaded into memory (for sparse cases).
This happens not only when we do vector search, but also when we open an index to 
load meta info for vector readers.

This patch instead uses IndexedDISI to store docIds and DirectMonotonicWriter/Reader to 
handle  ordToDoc mapping. Benefits are reduced memory usage, and faster loading of 
meta info for vector readers.