LUCENE-10560: Reduce the number of bytes compared while construction an OrdinalMap.
This changes OrdinalMap construction to work on windows of 4096 terms at a time
and identify shared prefixes within the window, so that the PriorityQueue can
ignore these shared prefixes when comparing its entries.

On the OrdinalMapBenchmark, this made things slightly slower because shared
prefixes within a window are only in the order of 2 bytes, so there are not
enough savings to amortize the additional overhead introduced by the seeks that
try to estimate shared prefixed within the next window.