LUCENE-10573: Improve stored fields bulk merge for degenerate O(n^2) merges.
With this change, if the first segment of the merge is too dirty, stored fields
would still perform a bulk merge until the first dirty chunk. In the degenerate
case when the first segment keeps being rewritten because it keeps being merged
with tiny segments, this greatly improves the performance of merging.

Before this change, luceneutil's StoredFieldsBenchmark runs in 78s on my
machine on 1M docs and `BEST_COMPRESSION`. After this change, it runs in 22s.
This is because merging a clean large segment with 9 small segments would
create a new large segment whose dirty chunks are located at the very end of
the segment. So the next merge could still bulk-copy most of the data.