LUCENE-10078: Enable merge-on-refresh by default.
This gives implementations of `findFullFlushMerges` to `LogMergePolicy` and
`TieredMergePolicy` and enables merge-on-refresh with a default timeout of
500ms.

The idea behind the 500ms default is that it felt both high-enough to have time
to run merges of small segments, and low enough that the freshness of the data
wouldn't look badly affected for users who have high refresh rates (e.g.
refreshing every second).

For `findFullFlushMerges`, `LogMergePolicy` looks at tail segments to see if it
can find at least `mergeFactor` flush segments below the min segment size, and
`TieredMergePolicy` looks for a merge that has at least `segmentsPerTier`
segments where the largest segment of the merge is a flush segment and below
the floor size.