Added triangulation monitor
The Monitor allows the triangulation algorithm to post informative
events to a listener.
The listener can be as simple as a counting listener to track events,
or as complex as something that analyses the complete state of the
triangulation.
For example the visualization monitor at https://github.com/craigtaverner/polygon-triangulator-debug
which creates PNG files for polygons, node lists and triangles,
for use in debugging or generating informative animations.