Introduce HNSW merge from graph prototype
Introduces prototype for initializing new graph during merge from an
existing graph. Graph selected for initialization must have 0 deleted
docs and have the comparatively the largest number of vectors in it.

In order to implement this, capability of inserting out of order nodes
into the graph had to be introduced. Additionally, scores for
initializer graph had to be recomputed.

Signed-off-by: John Mazanec <jmazane@amazon.com>