Prevent PointValues from returning null for ghost fields
getPointValues may currently return null for unknown fields or fields that don't index points.
It can happen that a field no longer has points for any document in a segment after delete+merge, which
causes field info to think that the field is there and has points, yet when calling getPointValues null
is returned.

With this change, we prevent getPointValues from returning null for ghost fields, it will instead return an
empty instance of PointValues.

Relates to #11393